feature/album-pagination/Album.js 모듈뭉치 분리
