+++ src/main/resources/templates/album/album-content.html
@@ -122,7 +122,7 @@
     </div>
 </section>
 <div layout:fragment="scripts">
-    <script type="module" src="/album/js/Album.js"></script>
+    <script type="module" src="/album/js/modules/albumMain.js"></script>
 </div>
 </body>
 </html>
